requirements.txt
aiogram==3.*
asyncpg==0.29.0
SQLAlchemy==2.0.30
python-dotenv==1.0.1
httpx==0.27.0
uvicorn==0.29.0
fastapi==0.111.0
pandas>=2.0.0
google-api-python-client>=2.154.0
google-auth>=2.36.0
google-auth-httplib2>=0.2.0